+++ OpenSim/Examples/ExampleLuxoMuscle/LuxoMuscle_create_and_simulate.cpp
@@ -282,12 +282,12 @@ int main(int argc, char* argv[]) {
             std::cin.get();
         }
     }
-    catch (OpenSim::Exception ex)
+    catch (const OpenSim::Exception& ex)
     {
         std::cout << ex.getMessage() << std::endl;
         return 1;
     }
-    catch (std::exception ex)
+    catch (const std::exception& ex)
     {
         std::cout << ex.what() << std::endl;
         return 1;